requirements.txt
flask
pyTelegramBotAPI
gtts
scikit-learn
pandas
python-dotenv
gunicorn
